Adhoc: Integer in entity objects replaced by int

+++ spring-batch-core/src/main/java/org/springframework/batch/core/step/NoWorkFoundStepExecutionListener.java
@@ -29,7 +29,7 @@ import org.springframework.batch.repeat.ExitStatus;
 public class NoWorkFoundStepExecutionListener extends StepExecutionListenerSupport {
 
 	public ExitStatus afterStep(StepExecution stepExecution) {
-		if (stepExecution.getItemCount().intValue() == 0) {
+		if (stepExecution.getItemCount() == 0) {
 			throw new NoWorkFoundException("Step has not processed any items");
 		}
 		return stepExecution.getExitStatus();
